# 8299807: String.newStringUTF8NoRepl and getBytesUTF8NoRepl always copy arrays