`JavaLangAccess::newStringUTF8NoRepl` and `JavaLangAccess::getBytesUTF8NoRepl` are not implemented correctly. They always copy arrays, rather than avoiding copying as much as possible as javadoc says.

I ran the tier1 test without any new errors.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [ ] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8299807](https://bugs.openjdk.org/browse/JDK-8299807): String.newStringUTF8NoRepl and getBytesUTF8NoRepl always copy arrays


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk pull/11897/head:pull/11897` \
`$ git checkout pull/11897`

Update a local copy of the PR: \
`$ git checkout pull/11897` \
`$ git pull https://git.openjdk.org/jdk pull/11897/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 11897`

View PR using the GUI difftool: \
`$ git pr show -t 11897`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/11897.diff">https://git.openjdk.org/jdk/pull/11897.diff</a>

</details>