@AlanBateman 2023-01-11
Would it be possible to provide some context on which public API you are testing with and the micro benchmark that you are using?