@stsypanov 2023-01-17
I see that `getBytesUTF8NoRepl()` is used in `j.u.z.ZipCoder`, so could you run tests for zip encoding just to make sure everything is fine there? And I think that tier2 should be run for this as well.