@RogerRiggs 2023-01-17
`newStringUTF*NoRepl` is not described or expected to optimizing the array allocation.
Nor does it mandate that callers must guarantee that the array is not re-used or modified.
The access to this method through JavaLangAccess is solely for the convenience of the algorithm.
I don't see the performance improvement as sufficient to offset the risk of mis-use and its possible loss of integrity.